+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to
 
 - 🔥(backend) remove mirroring feature
 
+### Fixed
+
+- 🐛(backend) fix indexing payload is_active and reach
+
 ## [v0.17.0] - 2026-04-23
 
 ### Added
@@ -36,7 +40,6 @@ and this project adheres to
 - 🐛(frontend) range selection freezes when there are many items in the list
 - 🐛(backend) fix openapi schema for item access endpoints
 - 🐛(backend) load jwks url when OIDC_RS_PRIVATE_KEY_STR is set
-
 
 ## [v0.16.0] - 2026-04-09